requirements.txt
soundfile
librosa
torchlibrosa
torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
laion-clap
# python-ffmpeg
ftfy
braceexpand
regex
sentence-transformers
transformers>=4.51.3
timm
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu124
decord
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu124
numpy>=1.24.2
scipy>=1.15.0
scikit-learn
pandas>=2.2.0
polars
h5py
tqdm
progressbar
llvmlite
webdataset
wget
